Add some fwb stubs from Oplus
This fixes Realme Link crash issue

Realme link tries to use reflection api for some oplus methods.
Normally the are not presents if it is not oppo phone. They catch exception and all work fine,
but on oplus/realme phones with imported oplus fwk for camera this doesn't work because we have
these methods in separate boot jar and got denial from system on these hidden api methods.
So, just move this harmless crap to FWB and annotate it properly.

Change-Id: I5ffecc9be52505371400b0b976c5f14e66bd853d