UdfpsHelper: Track brightness mirror to hide/show Udfps dim layer
This provides a much deeper integration with Udfps dim layer and
the new brightness slider such that we don't need to look for APIs
that can track temporary brightness. That is because when the
brightness mirror is shown, there is no necessity for the Udfps
dim layer to be present. This change also isn't functional with the
legacy slider.

It also fixes the issue where the dim layer is hidden with
QS expansion -> Device Control where the alternate bouncer is brought up.

Tests : Ensure that appropriate transparency of the dim layer is set
when dragging or selecting the preferred brightness with the slider.

Change-Id: I53f78cbcfd3b42a3b5dbe6110424687ebfac102d